c306: extract RawTab + SceneStats + SegmentationOverlayPanel + SegmentationOverlayRaster + SEGMENTATION_METHODS to own file (#441 2.1)
Workspace.tsx 5933 LOC -> 5535 LOC (-398 LOC). Entry chunk 155.88 KB -> 143.26 KB (-12.6 KB).
24 tabs extracted + lazy.
Removed newly-unused ClassDistributionBar + SpectralByClass imports from Workspace.tsx.
typecheck clean, 44/44 tests pass.
